requirements.txt
mysql-connector-python
psutil
pandas
pillow